+++ arms-20190808/core/Models/DeleteEnvPodMonitorResponse.cs
@@ -10,15 +10,15 @@ namespace AlibabaCloud.SDK.ARMS20190808.Models
 {
     public class DeleteEnvPodMonitorResponse : TeaModel {
         [NameInMap("headers")]
-        [Validation(Required=true)]
+        [Validation(Required=false)]
         public Dictionary<string, string> Headers { get; set; }
 
         [NameInMap("statusCode")]
-        [Validation(Required=true)]
+        [Validation(Required=false)]
         public int? StatusCode { get; set; }
 
         [NameInMap("body")]
-        [Validation(Required=true)]
+        [Validation(Required=false)]
         public DeleteEnvPodMonitorResponseBody Body { get; set; }
 
     }