+++ arms-20190808/core/Models/ListEnvironmentDashboardsRequest.cs
@@ -1,0 +1,42 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.ARMS20190808.Models
+{
+    public class ListEnvironmentDashboardsRequest : TeaModel {
+        /// <summary>
+        /// Name of Addon.AddonName and Scene are required.
+        /// </summary>
+        [NameInMap("AddonName")]
+        [Validation(Required=false)]
+        public string AddonName { get; set; }
+
+        /// <summary>
+        /// The ID of the environment instance.
+        /// </summary>
+        [NameInMap("EnvironmentId")]
+        [Validation(Required=false)]
+        public string EnvironmentId { get; set; }
+
+        /// <summary>
+        /// The ID of the region.
+        /// </summary>
+        [NameInMap("RegionId")]
+        [Validation(Required=false)]
+        public string RegionId { get; set; }
+
+        /// <summary>
+        /// Scene of Addon.AddonName and Scene are required
+        /// </summary>
+        [NameInMap("Scene")]
+        [Validation(Required=false)]
+        public string Scene { get; set; }
+
+    }
+
+}