+++ arms-20190808/core/Models/ListEnvironmentDashboardsResponse.cs
@@ -8,7 +8,7 @@ using Tea;
 
 namespace AlibabaCloud.SDK.ARMS20190808.Models
 {
-    public class ListEnvironmentsResponse : TeaModel {
+    public class ListEnvironmentDashboardsResponse : TeaModel {
         [NameInMap("headers")]
         [Validation(Required=false)]
         public Dictionary<string, string> Headers { get; set; }
@@ -19,7 +19,7 @@ namespace AlibabaCloud.SDK.ARMS20190808.Models
 
         [NameInMap("body")]
         [Validation(Required=false)]
-        public ListEnvironmentsResponseBody Body { get; set; }
+        public ListEnvironmentDashboardsResponseBody Body { get; set; }
 
     }